Feature/front do vini (#16)
* refactor: update property handling in FormPropertyController to use inspection date

* refactor: update property handling in user controllers to use RuralProperty and improve data loading

* refactor: update RuralProperty model and service methods to use Long for ID types and enhance property management

* refactor: remove address column from property tables in user views to streamline data presentation

+++ src/main/java/com/unip/repository/RuralPropertyRepository.java
@@ -14,10 +14,10 @@ import jakarta.transaction.Transactional;
 
 @Repository
 public interface RuralPropertyRepository extends JpaRepository<RuralProperty, Long> {
-    void deleteById(int id);
+    void deleteById(Long id);
 
     @Modifying
     @Transactional
     @Query("UPDATE RuralProperty r SET r.inspectionDate = :date WHERE r.id = :id")
-    void updateInspectionDate(@Param("id") int id, @Param("date") LocalDate date);
+    void updateInspectionDate(@Param("id") Long id, @Param("date") LocalDate date);
 }
